# Optimizing VRF computation by saving expanded private key and public key