--- a/akd_core/src/ecvrf/mod.rs
+++ b/akd_core/src/ecvrf/mod.rs
@@ -25,7 +25,9 @@
 mod ecvrf_impl;
 mod traits;
 // export the functionality we want visible
-pub use crate::ecvrf::ecvrf_impl::{Output, Proof, VRFPrivateKey, VRFPublicKey};
+pub use crate::ecvrf::ecvrf_impl::{
+    Output, Proof, VRFExpandedPrivateKey, VRFPrivateKey, VRFPublicKey,
+};
 pub use crate::ecvrf::traits::VRFKeyStorage;
 #[cfg(feature = "nostd")]
 use alloc::boxed::Box;