One more VRF computation optimization that we can do:

Rather than re-expanding the VRF private key into a VRFExpandedPrivateKey and recomputing the public key every time we call evaluate, we can just compute the expanded private key and public key once, and then reuse them across VRF evaluate calls.

Before this change:
```
Batch insertion (100000 initial leaves) took 14172 ms
Batch insertion (100000 initial leaves, 100000 inserted leaves) took 16920 ms
```

After this change:
```
Batch insertion (100000 initial leaves) took 13712 ms
Batch insertion (100000 initial leaves, 100000 inserted leaves) took 15315 ms
```